Clarify our wording around telemetry gathering (#4939)
* Clarify our wording around telemetry gathering

We need to clarify that we're collecting user anonymized usage data. As
tsmith logged into an Automate install no data will be collected that
will allow my personal usage to be determined. This wasn't entirely
clear and it's something we need to make crystal clear.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Update more telemetry wording

There were more places we needed to clarify the user nature

Signed-off-by: Tim Smith <tsmith@chef.io>