Remove environement column from Policygroup nodes
I've noticed this in a bunch of places in Automate and I think there's a
misunderstanding with how Policyfiles work. If you use Policyfiles you
can't use environments/roles. They are incompatible methods of managing
nodes. So if you're listing nodes that are managed with policyfiles you
don't car what their environment is.

Signed-off-by: Tim Smith <tsmith@chef.io>