Increase ProxySendTimeout for ES Gateway to prevent backup delete timeouts
When we delete an ES snapshot, we don't receive an `acknowledged` response
until the snapshot deletion completes. For systems with large amounts of data
this can easily exceed the current 10min deadline. This fix increases the
timeout to 2hrs. A large customer in EMEA has experience timeouts in the order
of 40mins for a single deletion.

Signed-off-by: Richard Nixon <richard.nixon@btinternet>